Move CLI Initializers
Our previous location for initializers was confusing and it was easy not
to know where they were. Instead, to make it more clear, let's move the
initializers to config/initalizers_cli.
