+++ bin/rmt-cli
@@ -12,9 +12,9 @@ require 'active_record'
 require 'erb'
 require 'yaml'
 require 'rmt/config'
-require 'initializers/disable_deprecation_warnings'
-require 'initializers/rmt_fast_gettext'
-require 'initializers/rmt_fast_gettext_cli_locale'
+require_relative '../config/initializers_cli/disable_deprecation_warnings'
+require_relative '../config/initializers_cli/rmt_fast_gettext'
+require_relative '../config/initializers_cli/rmt_fast_gettext_cli_locale'
 
 relative_load_paths = %w[lib lib/rmt app/models app/services app/validators].map { |dir| File.join(rmt_path, dir) }
 ActiveSupport::Dependencies.autoload_paths += relative_load_paths
