+++ config/initializers/fast_gettext.rb
@@ -1,4 +1,4 @@
-require 'initializers/rmt_fast_gettext'
+require_relative '../initializers_cli/rmt_fast_gettext'
 
 # enable comments in .po/.pot files prefixed with i18n tag (see rxgettext -h)
 Rails.application.config.gettext_i18n_rails.xgettext = %w[--add-comments=i18n]
