+++ config/initializers_cli/rmt_fast_gettext.rb
