+++ config/initializers_cli/rmt_fast_gettext_cli_locale.rb
